+++ static/main.jsx
@@ -9,7 +9,7 @@ import { DataViewer } from "./dtale/DataViewer";
 import { CodeExport } from "./popups/CodeExport";
 import { CodePopup } from "./popups/CodePopup";
 import { Correlations } from "./popups/Correlations";
-import { ReactDescribe as Describe } from "./popups/Describe";
+import { Describe } from "./popups/Describe";
 import { ReactFilter as Filter } from "./popups/Filter";
 import Instances from "./popups/Instances";
 import { ReactColumnAnalysis as ColumnAnalysis } from "./popups/analysis/ColumnAnalysis";